Creating a simple "Hello, World!" program
The "Hello, World!" is one of the most classic program in programming tutorials,
used by Brian Kernighan and Dennis Ritchie as the first example of the book
"The C Programming Language". Respecting the tradition, this is our first
functional C++ program.

To print a message in the console, we use the command `cout` (from `Console
OUTput`). The desired output is encoded as a **string** (a null-byte terminated
sequence of characters defined as a text between double quotes). `endl` is a
**manipulator** that says that `cout` must finish the message with a new line.
`<<` is the **output operator**, used by `cout` to receive strings and
manipulators.

Both `cout` and `endl` are declared in the header `iostream`. This header is
part of the **Standard Template Library** (STL) - a set of useful variables,
functions and classes available as part of C++. Input and Output (IO) are some
of the many useful operations. To make `cout` and `endl` available to your
program, we need to type `#include <iostream>`.

As part of the standard, everything in `iostream` is part of the `std`
namespace. **Namespaces** are a C++ feature that aims to avoid name clashes.
The importance of namespaces appear when using different libraries. In C and
C++, a symbol "foo" defined twice produces a compilation error. With
namespaces, we could have two "foo"'s if they were in different namespaces.
To access a symbol within a namespace, we use the double colon operator, `::`.
Because of that, in the program, we use `std::cout` and `std::endl`.

The expected output is: the text `Hello, World!` in the console.

Command list:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
Hello, World!
================================================================================