Defining variables and making simple arithmetic
In C++, variables are defined with a **type** (as `int`) and a name (as `a`).
They may be initialized using the attribution operator `=`, which must be
followed by an expression that evaluates to the value that will be stored
witin the variable. To represent signed integers, the C++ language counts with
the basic type `int`. Integers themselves can be written as a sequence of
numbers between 0 and 9, optionally started by a **sign** (characters `+`
or `-`), but never started with a left `0` (unless the integer is the number
zero itself). Each variable declaration goes in a separated line. In order
to simplify the creation of multiple variables of the same type, it's possible
to create a **variable list**, wich begins with the type and has a
comma-separated list of names, optionally initialized.

The basic integer binary operators are `+`, `-`, `*` and `/`, which represent
sum, subtraction, product and division. Their precedence follows the precedence
rules of simple arithmetic. Many other operators exist, but in this tutorial we
will focus only in this four. The complete list of C++ operators and their
precedences can be seen in: http://www.cplusplus.com/doc/tutorial/operators/

The expected output is: four numbers, representing the sum, subtraction, product
and division of `20` by `10`.

Command list:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
The result of `a + b` is: 30
The result of `a - b` is: 10
The result of `a * b` is: 200
The result of `a / b` is: 2
================================================================================