Reading values from the input
As a rule, any interesting application accepts data from its user, process it
and shows its results in some useful way. In C++, it is possible to read values
from the console using the command `cin` (from Console INput), defined in the
`std` namespace and available by the `iostream` header. `>>` is the **input
operator**, used by `cin` to point to the variables that will store the values
read. The type of the variable being read determine the format that has to be
written by the user in the console. In the case of signed integers, the format
written has to be the same of integer literals.

The expected output is: Two messages prompting the user for numbers to be used
in the sum, subtraction, product and division, then the result of executing
these operations. In the following examples, we will keep using `20` and `10`
as our standard input.

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
Input 1st number (a): 20
Input 2nd number (b): 10
The result of `a + b` is: 30
The result of `a - b` is: 10
The result of `a * b` is: 200
The result of `a / b` is: 2
================================================================================