Extracting common pieces of code
Repetition is one of the evils in programming. As humam memory is limited,
when we want to change repeated code, we can forget one of the snippets and
create inconsistences. To group and reuse useful pieces of code, we can create
a **function**, which is declared as follow: it contains a **return type**,
a **name** and an ordered set of **parameters**. In order to **call** the
function, we just need to put its name, and a comma separated list of
**arguments** enclosed by parenthesis. Arguments are just variables (`a`, `b`)
or literal values (`1`, `2`, `3`) that must have the same type as the
parameters. Within the function's body, parameters are nothing but variables,
whose values are filled with **copies** of the arguments. Parameter become
available within the body of the function, and can be used to make the
operations desired (as arithmetic sums, products, etc.). They can also be
applied as arguments to other function calls. The `return` clause is used
to pass some result (stored in a variable or resulted by an expression)
to the function's caller.

A function that just uses its parameters to produce a return value is
called **pure**, because its behavior is the same of a mathematical function.
However, when we read a value from the console, we are making a **collateral
effect** - an action that have a repercurssion outside the _scope_ of the body
of the functions. Collateral effects usually are related to IO (such as file,
network or database operations) and to modifications of **global variables**
(not seen yet).

The expected output is: The same as before. We just **refactored** our code
(made it clearer and more reusable, with no changes in the observable behavior).

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
Input 1st number (a): 20
Input 2nd number (b): 10
The result of `a + b` is: 30
The result of `a - b` is: 10
The result of `a * b` is: 200
The result of `a / b` is: 2
================================================================================