Deciding which operation to execute
It is ideal to allow the user to decide which arithmetic operation it wants to
execute. In our daily life (and in our C++ programs, so far), we use the
**infix notation** (number followed by an operator followed by a number) as our
standard way to write arithmetic expressions. Other representations do exist:
prefix notation (operator, number and number) and postfix notation (number,
number and operator) are used by some programming languages to write their
expressions. In order to continue our example, and to build a simple integer
calculator, we will try to read lines written in the infix notation.

Instead of a long input message, we prompt our user with a single character `>`,
which commonly indicates some space to type. The user can, then, write simple
expressions written with two numbers. The function `readNumber` is changed to
just read from the console, not outputing anything. A function `readCharacter`
does a similar task, reading a simple character.

In C++, character types can be represented in many ways. The simplest type,
`char`, is an 8 byte number, which can encode up to 256 different symbols.
A `char` variable is similar to a small integer, and can have positive and
negative values. The symbols from 0 to 127 are standardized in the ASCII table
(http://www.asciitable.com/), while the negative part (from -1 to -128) can
have different symbolic representations accordingly to the regional **encoding**
used (as latin-1). Characters can be represented both by their ASCII table
number or by single letters (typed in the keyboard) surrounded by quotes.
The function `cin` is smart enough to recognize the type of the variable
where it will put the console value read, and waits for the user to input
the right pattern. Therefore, when reading a variable, any character as `+`,
`-`, `*` or `/` will be accepted.

In order to identify which operation should be executed, we need some language
statement that allow us to _change the flow of execution_. For this task, we
have the **if...else statement**. It is defined as follow: the keyword `if`
is followed by an expression surrounded by parenthesis. This expression must
evaluate to a **boolean** (true/false) value. The simplest boolean expression
is an **equal** comparison, represented by the operator `==`. This operator
has two equal signs to desambiguate it from an attribution (`=`). `2 + 2 == 4`
evaluates to `true`. `2 + 2 == 5` evaluates to false. `a == b` checks if
the variables `a` and `b` have the same value. When an `if` condition evaluates
to `true`, the code within its body is executed. Otherwise, it is skipped.
When the `if` statement is followed by an `else`, the `else` body is executed
when the condition evalates to `false`.

Many if...else statements can be concatenated through an `else if (condition)`
written after the body of the first `if`. These language constructs are called
**if...else if...else statements**, and can be used to test many possibilities
**sequentially**. The first one to evaluate to `true` is the only one executed,
and the rest is skipped. If the statement has an `else` in its end, it will be
executed if none of the conditions evaluate to `true`. Otherwise, the whole
block is skipped. In the example, one of this blocks is used to decide if the
character variable `op` is one of the known operators. If not, an error message
is printed.

The expected output is: A simple prompt that accepts expressions between
two integers, and print a simple error message in case the operator is
unknown.

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of `a + b` is: 30
$ ./calculator
> 20 - 10
The result of `a - b` is: 10
$ ./calculator
> 20 * 10
The result of `a * b` is: 200
$ ./calculator
> 20 / 10
The result of `a / b` is: 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================