Fixing the output message
In our previous version, we left the original output message (referring to
`a` and `b`) behind. We can use the variable `op` with `cout`, which, as
`cin`, is smart enough to recognize the type of the variable (`char`) and
print the right ASCII character.

The expected output is: The same as before, but with a more meaningful
message representing the arithmetic expression executed.

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================