Improving code for formatted output
Print formatted ouput with `cout` can be a difficult (and verbose) task.
In order to simplify this kind of operation, we can use `printf`, a tipical
C solution for printing in the console. `printf` receives, as first parameter,
a **format string** with markers that specify the kind of value that will be
printed in the position of the mark. The format string can be very complex,
and it makes a complete DSL (domain specific language) by itself. To print
integers, we use `%d` (from "decimal") and for characters, we use `%c`
(from "character"). To make a line jump in the end, we put the mark `\n`,
which represents the character "end of line". Marks started by `%` get
substituted, as marks started by `\` represent "invisible" characters in
the ASCII table that change the way the text is printed.

After the format string, `printf` accepts a comma separated list of expressions
whose result value will be printed in the output. This expression can be made
from a single variable (`a`), an arithmetic expression (`a + b`) or even a
function call. Each argument is printed in the position of the format string
specified by the mark. Usually, compilers can warn when there is an error or
mismatch between the format string and its arguments.

`printf` is declared in the C header `stdio.h`. Although this header can
be used directly in C++ with an include directive (`#include <stdio.h>`),
it's the best practice to include `cstdio` insted (`#include <cstdio>`).
`cstidio` is a wrapper around `stdio.h` that declares each function available
in the `std` namespace. This way, it's possible to use `std::printf`, as in
`std::cout`.

As a last note, usually `printf` is faster than `cout` when printing the same
text. However, when time is not a critical issue, it's acceptable to choose
the one that will make printing easier.

The expected output is: The same as before. We just **refactored** our code
(made it clearer and more reusable, with no changes in the observable behavior).

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================