Separating calculation from output
Mixing the general program logic with the output of results is a bad practive.
When the output is replicated in an if-else block, as in our case, making
changes in the printed message is not as trivial as it could be. This kind of
replication is not new: it has already be seen with the input of data. The
primary solution in that situation was to extract the input in a function.
In this case, however, the output message is the same, while the selection
of the binary operation and its execution the variant code. Therefore,
instead of extracting a function for the output (as in the input case),
it is better to create a `calculate` function to hold the main operations
made by our calculator.

Instead of using a if...else if...else statement, we implement `calculate` with
a **switch-case statement**. For integral numeric types as `int` of characters
as `char`, this statements can be used as a less verbose way to decide which
option should be executed. In parenthesis, after the keyword `switch`, comes
a varaible or expression whose value will be compared to many possibilities,
all listed inside a block (code snippet between curly braces). Each possibility
is reprented by the keyword `case` followed by a number / character literal
(as `12345` or `'a'`) and a colon (`:`). The value beside `switch` is compared
to each literal. The first comparison to return `true` has the code following
it executed. However, instead of stopping in the next `case` (as would be with
an `else if` in a if...else if...else statement), the execution continues
until the end of the block. That is, all the code from the entry point until
the end of the switch-case statement is executed, ignoring all the comparisons
above. To avoid this, it is possible to use the keyword `break`, that interrupts
the execution and jumps directly to the end of the block. Using a `break` for
each `case` of a `switch` makes the switch-case statement behaves exactly as
a if...else if...else statement without an `else`. The `default` keyword,
followed by a `:`, simulates a case that always evaluates to `true`. Therefore,
everything after the default case gets executed, unless it has a `break`.
If the default case is used as the last possibility in a switch-case, this
`break` is unnecessary (as execution will stop in the following end of the
block) and its behavior is the same as the `else` of an if...else if...else
statement.

In our example, the switch-case statement maps directly an equivalent
if...else if... else statement. The variable `result` is initialized as `0`,
and this value is kept when the code enters in the default case.

The expected output is: The same as before. We just **refactored** our code
(made it clearer and more reusable, with no changes in the observable behavior).

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================