Declaring variables as late as possible
The smaller the scope of variables, the easier is to know where and when they
are necessary. In traditional (K&R and ANSI) C, variables could be declared
only in the beggining of a block (snippet of code between curly brackets).
Because of that, When functions grow, code become divided in two sections:
a list of variable declarations, and statements that actually do the function's
job. This division requires programmers to jump inside their editors whenever
they need to change the variables. In the long term, this extra effort slows
down programming and helps to introduce bugs (as programmers may try to reuse
their variables, and awkward things can happen with memory in these cases).

In C++ and modern versions of C, variables can be declared anywhere inside
a block. The rule of thumb is: **declare variables as late as possible**,
and always **initialize** it (attribute a value that will be used) when
making the declaration.

The expected output is: The same as before. We just **refactored** our code
(made it clearer and more reusable, with no changes in the observable behavior).

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================