Returning multiple values in a function
Functions can return different values accordingly to some conditions. Anywhere
in the body of a function, the `return` keyword can be used to stop its
execution. This is true even inside an if...else if... else or a switch-case
statement. Therefore, it is possible to simplify our example and inline the
variable `result`, returning directly the value of the arithmetic operations
to the caller of the function. A function with a return value, however, needs
to return something in all possible execution paths inside its body. Therefore,
the default case must end returning something. As there is no other option,
we arbitrarily choose `0` as our **error code**, although it cannot be
distinguished from the result of other arithmetic operations.

The expected output is: The same as before. We just **refactored** our code
(made it clearer and more reusable, with no changes in the observable behavior).

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================