Handling erros out of the program logic
Signaling errors for the user is hard. When the result of a function is limited
(as restricted to positive integers), it is possible to use a different range
of values (as negative integers) to warn users about errors. In C, this is the
default mechanism to communicate problems. However, it can also make programming
very hard. In the first place, there is no standardized way to say which results
mean errors (`0` is an error for most standard C functions, but in system calls
of Unix-based operating systems, it means the run was successful). In the second
place, whenever functions signal errors this way, the call must be followed by a
test, to guarantee everything worked properly. Mixing general program logic and
error handling makes the code more difficult to read, and is not ideal for
large systems with many functions.

C++ has a better approach to handle errors: **exceptions**, that are a typical
Object Oriented Programming (OOP) constrution. An exception is **throwed** when
it is created with the keyword `throw`. It is then propagated from function to
function, from the one who throwed to its callee, its callee callee, etc. If it
gets to `main`, the program aborts the execution.

STL provides a set of standard exceptions (available in the header `stdexcept`)
which represent the most common errors that a programmer may want to warn about.
In `calculate`, the default case represent that the character parameter `op`
does not represent a known operator. Therefore, it is an invalid argument for
the function. The exception `invalid_argument` express exactly this idea, and
it is the right choice for this function. The type `std::invalid_argument`
must come right after the keyword `throw`.

The last requirement to throw an exception is to define an **error message**,
to help the user to know what exactly happened. For simplicity, our first error
message is a simple string literal, with no reference to the wrong character
inserted (and passed to the function inside `op`). The error message comes
after the exception type between parenthesis. This way, the full syntax to
throw an exception is: `throw exception_type("error message");`.

The expected output is: The same as before, except in the case of the invalid
oeprator. The exception we throwed is received by `main`, that aborts the
program and prints the error message.

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
terminate called after throwing an instance of 'std::invalid_argument'
  what():  Operator is unknown
Aborted (core dumped)
================================================================================