Parameterizing functions by types
Encoding types in the name of functions is not ideal. Both `readNumber` and
`readCharacter` have almost the same code, differing only in the type of the
variables they are reading. In C, there is nothing better to do. As a
consequence, many C standard library functions have names that vary accordingly
to the type they accept, although they execute exactly the same code. This
have two major disadvantages: first, programmers have to rely on naming
conventions and documentation to explain which type each function accept
- a practice that can be hard to mantein; second, and as a consequence, users
loose productivity as they need to remember different naming conventions and
look for different documentations accordingly to the library they are using.

In C++, however, it is possible to parameterize functions using **templates**:
a language construction that creates "factories of code". In order to create a
template function, it is necessary to put the `template` with a comma-separated
list of typename declarations surrounded by angle brackets (`<` and `>`).
A **typename declaraion** is the keyword `typename` followed by an identifier,
that will be an inner alias for a type. This type can be used anywhere inside
the normal function definition: as its return value, as a type of one or many
parameters, or to declare variables inside the function's body. In our example,
`Type` is the type of the variable that will be read by `cin`, and returned
by `read`.

In order to call a template function, it is necessary to pass, between the
function's name and its argument list, a comma-separated list of types,
surrounded by angle brackets. For each combination of types in this list,
the compiler generates a specific function based on the function template.
When the function is called, it is not the template code that is being executed,
but the ones from these auto-generated versions.

As code generators, templates are a very powerful **metaprogramming** technique,
allowing new kinds of code reuse with no runtime (execution) cost. By their own,
they provide much of the generality and power of the C++ programming language.

The expected output is: The same as before. We just **refactored** our code
(made it clearer and more reusable, with no changes in the observable behavior).

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================