Deducing obvious types automatically
Writing types can be repetitive and error-prone. Sometimes, the name of the
type returned from a function is obvious from the function's name. With template
functions, this type may even be encoded direcly in the template argument list.
When creating variables from the returned type of this kind of functions,
writing the type twice creates replications (and, consequently, brings all its
evils). To avoid that, C++ has the `auto` keyword, that automatically deduces
(in compile time) the type of the variable from the result of the expression
in the right hand side of the attribution. If we change this expression, the
next compilation will keep the variable's name, but will change its type with
no need to change the code.

The expected output is: The same as before. We just **refactored** our code
(made it clearer and more reusable, with no changes in the observable behavior).

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
Error! Operator # is unknown
================================================================================