Explaining the course
Welcome to C++! An amazing language for high performance programming.

This course is aimed to explain, through a single example, many of the
most important features of C++. Every explanation is written in commits,
each introducing a small new concept of C++. The whole historic represents
the changes needed to make the final version of the example, and complete
this course.

At this moment, as I am writing the course, this repository will be very
unstable. Combining `git rebase -i` and `git push -f origin master` is not
a good practice, but it will be a necessary evil to write this course inside
the commits...

To read the course, use `git log -p` inside the terminal or access
https://github.com/renatocf/cpp-course/commits/master in a browser.

I hope you enjoy!

Renato Cordeiro Ferreira