Creating string stream objets to format error messages
The expected output is: The same as before, except in the case of the invalid
oeprator. With string streams, it is possible to exhibit the wrong operator
(stored in the character variable `op`) in the error message.

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
$ ./calculator
> 20 - 10
The result of 20 - 10 is 10
$ ./calculator
> 20 * 10
The result of 20 * 10 is 200
$ ./calculator
> 20 / 10
The result of 20 / 10 is 2
$ ./calculator
> 1 # 1
terminate called after throwing an instance of 'std::invalid_argument'
  what():  Operator # is unknown
Aborted (core dumped)
================================================================================