Accepting multiple queries in one run
The expected output is: A infinite loop shell that prompts for a binary
arithmetic expression. All the previous operations can be done, now, within
one run. The exception aborts the execution and ends the program.

Command lines:
================================================================================
$ CXX -std=c++14 calculator.cpp -o calculator
$ ./calculator
> 20 + 10
The result of 20 + 10 is 30
> 20 - 10
The result of 20 - 10 is 10
> 20 * 10
The result of 20 * 10 is 200
> 20 / 10
The result of 20 / 10 is 2
> 1 # 1
terminate called after throwing an instance of 'std::invalid_argument'
  what():  Operator # is unknown
Abortado (imagem do núcleo gravada)
================================================================================